replace shorten links services to url shortener ecosystem, update materials services
