--- Materials/Libraries/SciMaterials.Contracts.ShortLinks/SciMaterials.Contracts.ShortLinks.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\Common\SciMaterials.Contracts\SciMaterials.Contracts.csproj" />
-  </ItemGroup>
-
-</Project>
