--- Materials/Libraries/SciMaterials.Services.ShortLinks/SciMaterials.Services.ShortLinks.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Abstractions" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.5" />
-    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SciMaterials.Contracts.ShortLinks\SciMaterials.Contracts.ShortLinks.csproj" />
-    <ProjectReference Include="..\SciMaterials.Contracts.API\SciMaterials.Contracts.API.csproj" />
-    <ProjectReference Include="..\SciMaterials.DAL.Resources\SciMaterials.DAL.Resources.csproj" />
-    <ProjectReference Include="..\SciMaterials.Services\SciMaterials.Services.csproj" />
-  </ItemGroup>
-  
-</Project>
