+++ UrlShortener/SciMaterials.UrlShortener.Contracts/ILinkReplaceService.cs
@@ -1,0 +1,7 @@
+namespace SciMaterials.UrlShortener.Contracts;
+public interface ILinkReplaceService
+{
+	Task<string> RestoreLinksAsync(string text, CancellationToken Cancel = default);
+	Task<string> ShortenLinksAsync(string text, CancellationToken Cancel = default);
+}
+
