+++ UrlShortener/SciMaterials.UrlShortener.Contracts/ILinkShortCutService.cs
@@ -1,0 +1,10 @@
+﻿using SciMaterials.Contracts.Result;
+
+namespace SciMaterials.UrlShortener.Contracts;
+
+public interface ILinkShortCutService
+{
+	Task<Result<string>> AddAsync(string sourceAddress, CancellationToken Cancel = default);
+	Task<Result<string>> GetAsync(string sourceAddress, bool isRedirect = false, CancellationToken Cancel = default);
+	string GetLinkBasePath();
+}
