+++ UrlShortener/SciMaterials.UrlShortener.Contracts/IUrlShortenerClient.cs
@@ -5,6 +5,6 @@ public interface IUrlShortenerClient
 {
 	string ShortenedUrlBase { get; }
 
-	Task<Result<string>> GetOriginalUrlAsync(string url);
-	Task<Result<string>> ShortenAsync(string originalUrl);
+	Task<Result<string>> GetOriginalUrlAsync(string url, CancellationToken cancellationToken);
+	Task<Result<string>> ShortenAsync(string originalUrl, CancellationToken cancellationToken);
 }
